+++ client/src/smIndex.ts
@@ -1,16 +1,37 @@
 import * as vscode from "vscode";
 import { registerSMLinter } from "./smLinter";
+import * as glob from "glob";
 import { SM_MODE } from "./smMode";
 import { Providers } from "./Providers/smProviders";
 import { registerSMCommands } from "./Commands/registerCommands"; 
 
 
+let getDirectories = function (src, callback) {
+  glob(src + '/**/*.sp', callback);
+};
+
+
 export function activate(context: vscode.ExtensionContext) {
   let providers = new Providers(context.globalState);
   
-	let sm_home : string = vscode.workspace.getConfiguration("sourcepawnLanguageServer").get(
+  // Parse files at document opening.
+  let sm_home : string = vscode.workspace.getConfiguration("sourcepawnLanguageServer").get(
 		"sourcemod_home");
   providers.parse_sm_api(sm_home);
+  let workspace : vscode.WorkspaceFolder = vscode.workspace.workspaceFolders[0];
+  if(typeof workspace != "undefined")
+  {
+    getDirectories(workspace.uri.fsPath, function (err, res) {
+      if (err) {
+        console.log("Couldn't read .sp file, ignoring : ", err);
+      } else {
+        for(let file of res)
+        {
+          providers.handle_document_opening(file);
+        }
+      }
+    });
+  }
 
   context.subscriptions.push(providers.completionsProvider);
 	context.subscriptions.push(vscode.languages.registerCompletionItemProvider(SM_MODE , providers.completionsProvider));